Add a message with document when auto accept a psy
