+++ package.json
@@ -48,6 +48,7 @@
     "compression": "^1.7.4",
     "cors": "^2.8.5",
     "cron": "^1.8.2",
+    "crypto": "^1.0.1",
     "dotenv": "^10.0.0",
     "ejs": "^3.1.5",
     "express": "^4.17.1",
@@ -64,6 +65,7 @@
     "intl": "^1.2.5",
     "jsonwebtoken": "^8.5.1",
     "knex": "^0.21.17",
+    "mime-types": "^2.1.31",
     "nodemailer": "^6.4.18",
     "pg": "^8.5.1",
     "ts-node": "^10.0.0",
