+++ test/demarcheSimplifiees/autoVerifyPsychologist.spec.js
@@ -2,7 +2,7 @@ const rewire = require('rewire');
 const sinon = require('sinon');
 const config = require('../../utils/config');
 
-const demarchesSimplifiees = rewire('../../services/demarchesSimplifiees');
+const autoVerify = rewire('../../services/demarchesSimplifiees/autoVerify');
 const graphql = rewire('../../utils/graphql');
 
 describe('autoVerifyPsychologist', () => {
@@ -13,7 +13,7 @@ describe('autoVerifyPsychologist', () => {
     unsets = [];
     executeMutationStub = sinon.stub();
     unsets.push(graphql.__set__('executeMutation', executeMutationStub));
-    unsets.push(demarchesSimplifiees.__set__('graphql', graphql));
+    unsets.push(autoVerify.__set__('graphql', graphql));
   });
 
   afterEach((done) => {
@@ -22,7 +22,7 @@ describe('autoVerifyPsychologist', () => {
   });
 
   it('Should execute 4 mutations on DS', async () => {
-    await demarchesSimplifiees.autoVerifyPsychologist();
+    await autoVerify.autoVerifyPsychologist();
 
     sinon.assert.callCount(executeMutationStub, 4);
